fix(security)+refactor: OAuth state invalidation, YAML path allowlist, server.py extraction, agent_remediation split
Refs #354 #355 #360 #361

fix(security): delete OAuth state after single use; add org-membership check via GitHub API (#354)
- routers/auth.py: del request.session["oauth_state"] immediately after state validation
- routers/auth.py: if REQUIRED_GITHUB_ORG is set, call /orgs/{org}/members/{login} and 403
  if the response is not 204 (non-member)

fix(security): validate AGENT_REMEDIATION_CONFIG env-var paths against allowed roots (#355)
- agent_remediation/policy.py: _validate_policy_path() checks resolved path is within
  <repo>/config/ or ~/.config/runner-dashboard before load_policy/save_policy proceed
- Prevents path traversal if an attacker controls the environment variable

refactor(server): extract repos + diagnostics routers from server.py (#360)
- backend/routers/repos.py: GET /api/repos, /api/prs, /api/issues,
  /api/tests/ci-results, POST /api/tests/rerun, GET /api/stats, /api/usage
- backend/routers/diagnostics.py: GET /api/diagnostics/summary,
  POST /api/diagnostics/restart-service, POST /api/launchers/generate,
  GET/POST /api/runner-routing-audit[/refresh]
- server.py: old handlers stubbed with comments; new routers registered via
  app.include_router; deps wired in _startup()

refactor(agent_remediation): split 934-line module into providers/policy/planner (#361)
- agent_remediation/providers.py: AgentProvider, ProviderAvailability, PROVIDERS registry,
  probe_provider_availability()
- agent_remediation/policy.py: data classes (RemediationPolicy, FailureContext, etc.),
  load_policy/save_policy, classify_workflow_type, build_failure_fingerprint
- agent_remediation/planner.py: plan_dispatch, provider_prompt, inspect_jules_workflows,
  DispatchDecision, WorkflowHealthReport
- agent_remediation/__init__.py: re-exports all public names for backwards compatibility
